requirements.txt
pytrends>=4.9.2
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
plotly>=5.18.0
streamlit>=1.30.0
openpyxl>=3.1.2
apscheduler>=3.10.4
pyyaml>=6.0.1
requests>=2.31.0
